+++ src/main/java/com/ace/ng/dispatch/javassit/CmdHandlerCreator.java
@@ -3,5 +3,5 @@ package com.ace.ng.dispatch.javassit;
 import com.ace.ng.dispatch.message.CmdHandler;
 
 public interface CmdHandlerCreator {
-	CmdHandler<?> create(Short cmd);
+	CmdHandler<?> create(Integer cmd);
 }
